added documentation to classes reverted token expiration to be given in seconds until expiration for ensuring using the server-time

+++ backend/src/main/java/communication/packets/response/LoginResponsePacket.java
@@ -2,15 +2,25 @@ package communication.packets.response;
 
 import communication.packets.PacketType;
 import communication.packets.ResponsePacket;
+import communication.packets.request.LoginRequestPacket;
 import login.LoginResponse;
 
+/**
+ * A login response sent as result of an login request using an {@link LoginRequestPacket}.
+ */
 public class LoginResponsePacket extends ResponsePacket {
 
     private LoginResponse loginResponse;
     private String token;
-    private long expiration;
+    private int expiration;
 
-    public LoginResponsePacket(LoginResponse loginResponse, String token, long expiration) {
+    /**
+     *
+     * @param loginResponse A String representing the overall result of the login request. This {@link LoginResponse} may be an abstraction of the actual response.
+     * @param token A token used for further communication on a successful login, null otherwise.
+     * @param expiration The number of seconds until the token should expire.
+     */
+    public LoginResponsePacket(LoginResponse loginResponse, String token, int expiration) {
         super(PacketType.LOGIN_RESPONSE);
         this.loginResponse = loginResponse;
         this.token = token;
